% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{example_filename}
\alias{example_filename}
\title{Retrieve an example filename}
\usage{
example_filename(
  what = c("template", "seguin_whitney2022", "lower-spies2020")[2]
)
}
\arguments{
\item{what}{char, one of 'template' or 'seguin_whitney2022' or 'lower-spies2020'}
}
\value{
character URL
}
\description{
Retrieve an example filename
}
