% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wds-class.R
\name{WDS_paleo}
\alias{WDS_paleo}
\title{Create a WDSpaleo object.}
\usage{
WDS_paleo(filename = example_filename("seguin_whitney2022"))
}
\arguments{
\item{filename}{char, the URL or filename to read}
}
\value{
WDSpaleo object
}
\description{
Create a WDSpaleo object.
}
\examples{
\dontrun{
X = WDS_paleo(filename = example_filename("seguin_whitney2022"))
}
}
