% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wds-class.R
\name{WDSpaleo}
\alias{WDSpaleo}
\title{WDSpaleo to access NOAA template files}
\description{
WDSpaleo to access NOAA template files

WDSpaleo to access NOAA template files
}
\details{
R6 class for accessing WDS-paelo template format
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{char, the URL (generally) of filename}

\item{\code{raw}}{char, character vector of raw data by line, don't mess with this}

\item{\code{data}}{tibble of data (or NULL)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WDSpaleo-new}{\code{WDSpaleo$new()}}
\item \href{#method-WDSpaleo-find_token}{\code{WDSpaleo$find_token()}}
\item \href{#method-WDSpaleo-parse_data}{\code{WDSpaleo$parse_data()}}
\item \href{#method-WDSpaleo-read_file}{\code{WDSpaleo$read_file()}}
\item \href{#method-WDSpaleo-clone}{\code{WDSpaleo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WDSpaleo-new"></a>}}
\if{latex}{\out{\hypertarget{method-WDSpaleo-new}{}}}
\subsection{Method \code{new()}}{
Create a new WDSpaleo instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WDSpaleo$new(filename = example_filename("seguin_whitney2022"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{char, the URL or filename to read}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WDSpaleo-find_token"></a>}}
\if{latex}{\out{\hypertarget{method-WDSpaleo-find_token}{}}}
\subsection{Method \code{find_token()}}{
Read the file storing either raw character data or NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WDSpaleo$find_token(token = "# Data:", fixed = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{char, the pattern to look for}

\item{\code{fixed}}{logical, passed to \code{\link[base]{grep}}}

\item{\code{...}}{other arguments for \code{\link[base]{grep}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
logical, TRUE if successful
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WDSpaleo-parse_data"></a>}}
\if{latex}{\out{\hypertarget{method-WDSpaleo-parse_data}{}}}
\subsection{Method \code{parse_data()}}{
Read the file storing either raw character data or NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WDSpaleo$parse_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
WDSpaleo object (suitable for piping)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WDSpaleo-read_file"></a>}}
\if{latex}{\out{\hypertarget{method-WDSpaleo-read_file}{}}}
\subsection{Method \code{read_file()}}{
Read the file storing either raw character data or NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WDSpaleo$read_file()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
logical, TRUE if successful
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WDSpaleo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WDSpaleo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WDSpaleo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
